Adding FIFO mediate shell wrapper and entry.
This mechanism allows an alias like this:

    alias box=". /path/to/bin/box"

To execute commands like this:

    box :box-add myproj

The purose being to allow the initiating process to run recieved
commands, in particular directory changes and also environment
variable setting. This is particularly important for making the
calling process vacate directories that want to be deleted but would
otherwise be blocked.