Fix issue where we only ignored a single parent of a merge.
The syntax "^thisref^" will omit the parent and any ancestors of
"thisref". If this is a merge commit, it only excludes the first
parent (and ancestors).

It would seem, according to gitrevisions(7), that the proper way to do
this is "thisref^!" which excludes all parents and their ancestors.
However, this seems to cause other problems. The righter solution is
apparently "^thisref^@"... which /actually/ excludes all parents and
their ancestors.

Frankly, git can go @#$#^&$%^thisref...