New 'marked-newest-ver-by-spec querying code.
This is not yet hooked into the CLI yet but is placed as development
alternate.

This new 'marked-newest-ver-by-spec querying code is believed to be
correct. It has been tested against all previously known degenerate
cases and handles them properly. There are also a couple of other
unexpected cases that we have discovered which it also handles
properly.

This code is not superbly fast (taking nearly 3s for our largest repo)
but it is usably fast.

The naming is terrible but will be cleaned up on refactor. Names are
largely designated to indicate what phase (listed in the docstring) is
being performed.