Incremental voomdb updates were done incorrectly, creating corrupt data
Our use of git fast-export caused the oldest commit(s) imported to claim a
change for every file in the repo. We now use import-marks to stop that
behavior.

Note the db header bump is not because the structure of the db changed, but
rather to trigger the removal of old and possibly corrupt dbs.