Fixing long-standing recursive build-deps bug.
Calls to build-deps may build projects using the same working copy but
depending on a different checkout state. Because the checkout state
was set prior to the call to build-deps, the checkout state may have
changed by install time. This didn't result in bad artifacts, our
safety mechanisms assure the proper version is being built, however it
did stop the build with an exception.

The previous work-around was to run "lein voom build-deps" enough
times, making a bit more progress each time, eventually succeeding.

This is no longer needed. Builds should run through to completion the
first time.

The solution is to set the checkout state before build-deps (as was
done previously) but to set the checkout state again, after
build-deps.